Title: 낮은 지점들, Time: 48ms, Memory: 2MB, Status: Passed - CodeTree

+++ 240203/낮은 지점들/lowest-points.cpp
@@ -3,17 +3,17 @@
 using namespace std;
 
 int n;
-unordered_map<int, int> Coord;
+unordered_map<long long, long long> Coord;
 
 int main() {
 	ios::sync_with_stdio(false); cin.tie(0); cout.tie(0);
 	cin >> n;
 	for (int i = 0; i < n; i++) {
-		int x, y; cin >> x >> y;
+		long long x, y; cin >> x >> y;
 		if (Coord.find(x) == Coord.end() || Coord[x] > y) Coord[x] = y;
 	}
 	
-	int ans = 0;
+	long long ans = 0;
 	for (auto it = Coord.begin(); it != Coord.end(); it++) {
 		ans += it->second;
 	}
